requirements.txt
Django==1.4.3
South==0.7.6
distribute==0.6.28
dj-database-url==0.2.1
django-registration==0.8

matplotlib==1.2.0
psycopg2==2.4.6
python-memcached==1.48
wsgiref==0.1.2
simplejson==3.0.7
